// calendar-app/src/app/services/models/GameDateComponent.ts
// Interface automatically generated by schemas-to-ts

export enum Season {
  SPRING = 'spring',
  SUMMER = 'summer',
  FALL = 'fall',
  WINTER = 'winter',
}

export interface GameDateComponent {
  id: string;
  season: Season;
  day: number;
  year?: number;
  isRecurring: boolean;
}
export interface GameDateComponent_Plain {
  season: Season;
  day: number;
  year?: number;
  isRecurring: boolean;
}

export interface GameDateComponent_NoRelations {
  season: Season;
  day: number;
  year?: number;
  isRecurring: boolean;
}